Split cache handlers into cache_handler.py
